cập nhật cấu trúc component từ PatientDetail.tsx sang MedicalManagement.tsx
