+++ src/pages/doctor/PatientDetail.tsx
@@ -1,22 +1,20 @@
 import { useEffect, useState } from "react";
-import { Typography, Row, Col, Tabs, Form, message } from "antd";
-import { MedicalDetailModal } from "./components/modals/MedicalDetailModal";
-import { TestResultModal } from "./components/modals/TestResultModal";
+import { Typography, Row, Col, Tabs } from "antd";
+// import { TestResultModal } from "./components/modals/TestResultModal";
 // import { TreatmentRoadMapModal } from "./components/modals/TreatmentRoadMapModal";
 import { PatientInformation } from "./PatientInformation";
 import { DoctorSidebar } from "./DoctorSidebar";
-import dayjs from "dayjs";
 import { MedicalManagement } from "./MedicalManagement";
 import DoctorApi from "../../servers/doctor.api";
 import { useParams } from "react-router-dom";
 import type { InformationPatientDetails } from "../../types/doctor";
 import type {
-  ConsulationResult_typeTest,
+  // ConsulationResult_typeTest,
   MedicalRecord,
-  MedicalRecordDetail,
-  TreatmentResult_typeTest,
-  treatmentRoadmap,
-  TypeTest,
+  // MedicalRecordDetail,
+  // TreatmentResult_typeTest,
+  // treatmentRoadmap,
+  // TypeTest,
   // UpdateTreatmentResultFormValues,
 } from "../../types/medicalRecord.d";
 // import { TreatmentResultModal } from "./components/modals/TreatmentResultModal";
@@ -24,10 +22,10 @@ import { MedicalRecordOverview } from "./MedicalRecordOverview";
 
 const { Title, Text } = Typography;
 
-interface TestResultFormValues {
-  name: string;
-  description: string;
-}
+// interface TestResultFormValues {
+//   name: string;
+//   description: string;
+// }
 
 export default function PatientDetailPage() {
   const { customerId } = useParams<{ customerId: string }>();
@@ -44,35 +42,36 @@ export default function PatientDetailPage() {
 
   const [medicalRecord, setMedicalRecord] = useState<MedicalRecord[]>([]);
 
-  const [treatmentRoadmap, setTreatmentRoadmap] = useState<treatmentRoadmap[]>(
-    []
-  );
-  const [medicalRecordDetails, setMedicalRecordDetails] = useState<
-    MedicalRecordDetail[]
-  >([]);
+  // const [treatmentRoadmap, setTreatmentRoadmap] = useState<treatmentRoadmap[]>(
+  //   []
+  // );
 
-  const [treatmentResult_typeTest, settreatmentResult_typeTest] = useState<
-    TreatmentResult_typeTest[]
-  >([]);
+  // const [medicalRecordDetails, setMedicalRecordDetails] = useState<
+  //   MedicalRecordDetail[]
+  // >([]);
 
-  const [consulationResult_typeTest, setConsultationResult_TypeTest] = useState<
-    ConsulationResult_typeTest[]
-  >([]);
+  // const [treatmentResult_typeTest, settreatmentResult_typeTest] = useState<
+  //   TreatmentResult_typeTest[]
+  // >([]);
 
-  const [testResults, setTestResults] = useState<TypeTest[]>([]);
+  // const [consulationResult_typeTest, setConsultationResult_TypeTest] = useState<
+  //   ConsulationResult_typeTest[]
+  // >([]);
+
+  // const [testResults, setTestResults] = useState<TypeTest[]>([]);
   const [activeTab, setActiveTab] = useState("overview");
-  const [isTestModalVisible, setIsTestModalVisible] = useState(false);
+  // const [isTestModalVisible, setIsTestModalVisible] = useState(false);
 
   //medical record detail
-  const [
-    isUpdateMedicalDetailModalVisible,
-    setIsUpdateMedicalDetailModalVisible,
-  ] = useState(false);
+  // const [
+  //   isUpdateMedicalDetailModalVisible,
+  //   setIsUpdateMedicalDetailModalVisible,
+  // ] = useState(false);
 
-  const [
-    isCreateMedicalDetailModalVisible,
-    setIsCreateMedicalDetailModalVisible,
-  ] = useState(false);
+  // const [
+  //   isCreateMedicalDetailModalVisible,
+  //   setIsCreateMedicalDetailModalVisible,
+  // ] = useState(false);
 
   //treatment result - type test
   // const [
@@ -83,13 +82,13 @@ export default function PatientDetailPage() {
   // const [editingTreatmentResult, setEditingTreatmentResult] =
   //   useState<TreatmentResult_typeTest | null>(null);
 
-  const [treatmentResultForm] = Form.useForm();
+  // const [treatmentResultForm] = Form.useForm();
 
   //medical record detail
-  const [editingMedicalDetail, setEditingMedicalDetail] =
-    useState<MedicalRecordDetail | null>(null);
+  // const [editingMedicalDetail, setEditingMedicalDetail] =
+  //   useState<MedicalRecordDetail | null>(null);
 
-  const [medicalDetailForm] = Form.useForm();
+  // const [medicalDetailForm] = Form.useForm();
 
   useEffect(() => {
     const fetchInforPatient = async () => {
@@ -101,25 +100,25 @@ export default function PatientDetailPage() {
     fetchInforPatient();
   }, [customerIdNumber]);
 
-  useEffect(() => {
-    const fetchTreatmentRoadmap = async () => {
-      if (customerIdNumber != null) {
-        const res = await DoctorApi.GetTreatmentRoadmap1(customerIdNumber);
-        setTreatmentRoadmap(res.data);
-      }
-    };
-    fetchTreatmentRoadmap();
-  }, [customerIdNumber]);
+  // useEffect(() => {
+  //   const fetchTreatmentRoadmap = async () => {
+  //     if (customerIdNumber != null) {
+  //       const res = await DoctorApi.GetTreatmentRoadmap1(customerIdNumber);
+  //       setTreatmentRoadmap(res.data);
+  //     }
+  //   };
+  //   fetchTreatmentRoadmap();
+  // }, [customerIdNumber]);
 
-  useEffect(() => {
-    const fetchMedicalRecordDetail = async () => {
-      if (customerIdNumber != null) {
-        const res = await DoctorApi.GetMedicalRecordDetails(customerIdNumber);
-        setMedicalRecordDetails(res.data);
-      }
-    };
-    fetchMedicalRecordDetail();
-  }, [customerIdNumber]);
+  // useEffect(() => {
+  //   const fetchMedicalRecordDetail = async () => {
+  //     if (customerIdNumber != null) {
+  //       const res = await DoctorApi.GetMedicalRecordDetails(customerIdNumber);
+  //       setMedicalRecordDetails(res.data);
+  //     }
+  //   };
+  //   fetchMedicalRecordDetail();
+  // }, [customerIdNumber]);
 
   // useEffect(() => {
   //   const fetchTreatmentResult_Typetest = async () => {
@@ -137,18 +136,18 @@ export default function PatientDetailPage() {
   //   fetchTreatmentResult_Typetest();
   // }, [customerIdNumber]);
 
-  useEffect(() => {
-    const fetchConsulationR_TestT = async () => {
-      if (customerIdNumber != null) {
-        const res = await DoctorApi.GetConsultaionResult_TypeTest(
-          customerIdNumber
-        );
-        console.log("Data fetchConsulationR_TestT:", res.data);
-        setConsultationResult_TypeTest(res.data);
-      }
-    };
-    fetchConsulationR_TestT();
-  }, [customerIdNumber]);
+  // useEffect(() => {
+  //   const fetchConsulationR_TestT = async () => {
+  //     if (customerIdNumber != null) {
+  //       const res = await DoctorApi.GetConsultaionResult_TypeTest(
+  //         customerIdNumber
+  //       );
+  //       console.log("Data fetchConsulationR_TestT:", res.data);
+  //       setConsultationResult_TypeTest(res.data);
+  //     }
+  //   };
+  //   fetchConsulationR_TestT();
+  // }, [customerIdNumber]);
 
   //Lấy danh sách medical record
   useEffect(() => {
@@ -166,22 +165,22 @@ export default function PatientDetailPage() {
   }, [customerIdNumber]);
 
   //Medical record details
-  const showUpdateMedicalDetailModal = (medicalDetail: MedicalRecordDetail) => {
-    console.log("Clicked medical detail:", medicalDetail);
-    setEditingMedicalDetail(medicalDetail);
-    setIsUpdateMedicalDetailModalVisible(true);
-    medicalDetailForm.setFieldsValue({
-      date: dayjs(medicalDetail.date),
-      testResult: medicalDetail.testResult || "",
-      note: medicalDetail.note || "",
-      status: medicalDetail.status,
-    });
-  };
+  // const showUpdateMedicalDetailModal = (medicalDetail: MedicalRecordDetail) => {
+  //   console.log("Clicked medical detail:", medicalDetail);
+  //   setEditingMedicalDetail(medicalDetail);
+  //   setIsUpdateMedicalDetailModalVisible(true);
+  //   medicalDetailForm.setFieldsValue({
+  //     date: dayjs(medicalDetail.date),
+  //     testResult: medicalDetail.testResult || "",
+  //     note: medicalDetail.note || "",
+  //     status: medicalDetail.status,
+  //   });
+  // };
 
-  const showCreateMedicalDetailModal = () => {
-    setIsCreateMedicalDetailModalVisible(true);
-    medicalDetailForm.resetFields();
-  };
+  // const showCreateMedicalDetailModal = () => {
+  //   setIsCreateMedicalDetailModalVisible(true);
+  //   medicalDetailForm.resetFields();
+  // };
 
   // const showUpdateTreatmentResultModal = (record: TreatmentResult_typeTest) => {
   //   setEditingTreatmentResult(record);
@@ -277,116 +276,117 @@ export default function PatientDetailPage() {
 
   //Medical record detail
 
-  const handleUpdateMedicalRecordDetail = async (updated: {
-    date: string;
-    testResult?: string;
-    note?: string;
-    status: string;
-  }) => {
-    if (!customerIdNumber || !editingMedicalDetail) return;
+  // const handleUpdateMedicalRecordDetail = async (updated: {
+  //   date: string;
+  //   testResult?: string;
+  //   note?: string;
+  //   status: string;
+  // }) => {
+  //   if (!customerIdNumber || !editingMedicalDetail) return;
 
-    try {
-      await DoctorApi.UpdateMedicalRecordDetail(
-        customerIdNumber,
-        editingMedicalDetail.medicalRecordDetailId,
-        {
-          date: updated.date,
-          testResult: updated.testResult || "",
-          note: updated.note || "",
-          status: updated.status,
-        }
-      );
+  //   try {
+  //     await DoctorApi.UpdateMedicalRecordDetail(
+  //       customerIdNumber,
+  //       editingMedicalDetail.medicalRecordDetailId,
+  //       {
+  //         date: updated.date,
+  //         testResult: updated.testResult || "",
+  //         note: updated.note || "",
+  //         status: updated.status,
+  //       }
+  //     );
 
-      // Cập nhật state
-      setMedicalRecordDetails((prev) =>
-        prev.map((item) =>
-          item.medicalRecordDetailId ===
-          editingMedicalDetail.medicalRecordDetailId
-            ? { ...item, ...updated }
-            : item
-        )
-      );
+  //     // Cập nhật state
+  //     setMedicalRecordDetails((prev) =>
+  //       prev.map((item) =>
+  //         item.medicalRecordDetailId ===
+  //         editingMedicalDetail.medicalRecordDetailId
+  //           ? { ...item, ...updated }
+  //           : item
+  //       )
+  //     );
 
-      message.success("Cập nhật hồ sơ bệnh án thành công");
-      setIsUpdateMedicalDetailModalVisible(false);
-    } catch (error) {
-      console.error("Lỗi khi cập nhật hồ sơ bệnh án:", error);
-      message.error("Cập nhật thất bại");
-    }
-  };
+  //     message.success("Cập nhật hồ sơ bệnh án thành công");
+  //     setIsUpdateMedicalDetailModalVisible(false);
+  //   } catch (error) {
+  //     console.error("Lỗi khi cập nhật hồ sơ bệnh án:", error);
+  //     message.error("Cập nhật thất bại");
+  //   }
+  // };
 
-  const handleCreateMedicalRecordDetail = async (values: {
-    treatmentRoadmapId: number;
-    date: string;
-    typeName: string;
-    testResult?: string;
-    note?: string;
-    status: string;
-  }) => {
-    if (!customerIdNumber) return;
+  // const handleCreateMedicalRecordDetail = async (values: {
+  //   treatmentRoadmapId: number;
+  //   date: string;
+  //   typeName: string;
+  //   testResult?: string;
+  //   note?: string;
+  //   status: string;
+  // }) => {
+  //   if (!customerIdNumber) return;
 
-    try {
-      const roadmapId = Number(values.treatmentRoadmapId);
-      const response = await DoctorApi.CreateMedicalRecordDetail(
-        customerIdNumber,
-        {
-          treatmentRoadmapId: roadmapId,
-          date: values.date,
-          typeName: values.typeName,
-          testResult: values.testResult || "",
-          note: values.note || "",
-          status: values.status,
-        }
-      );
+  //   try {
+  //     const roadmapId = Number(values.treatmentRoadmapId);
+  //     const response = await DoctorApi.CreateMedicalRecordDetail(
+  //       customerIdNumber,
+  //       {
+  //         treatmentRoadmapId: roadmapId,
+  //         date: values.date,
+  //         typeName: values.typeName,
+  //         testResult: values.testResult || "",
+  //         note: values.note || "",
+  //         status: values.status,
+  //       }
+  //     );
 
-      // Tìm roadmap tương ứng để lấy stage và stepNumber
-      const relatedRoadmap = treatmentRoadmap.find(
-        (r) => r.treatmentRoadmapId === roadmapId
-      );
+  //     // Tìm roadmap tương ứng để lấy stage và stepNumber
+  //     const relatedRoadmap = treatmentRoadmap.find(
+  //       (r) => r.treatmentRoadmapId === roadmapId
+  //     );
 
-      setMedicalRecordDetails((prev) => [
-        ...prev,
-        {
-          medicalRecordDetailId: response.data.medicalRecordDetailId,
-          treatmentRoadmapId: roadmapId,
-          date: dayjs(values.date).format("YYYY-MM-DD"),
-          typeName: values.typeName,
-          testResult: values.testResult || "",
-          note: values.note || "",
-          status: values.status,
-          stage: relatedRoadmap?.stage || "", // Cung cấp giá trị mặc định nếu cần
-          stepNumber: relatedRoadmap?.stepNumber || 0, // Cung cấp giá trị mặc định nếu cần
-        },
-      ]);
+  //     setMedicalRecordDetails((prev) => [
+  //       ...prev,
+  //       {
+  //         medicalRecordDetailId: response.data.medicalRecordDetailId,
+  //         treatmentRoadmapId: roadmapId,
+  //         date: dayjs(values.date).format("YYYY-MM-DD"),
+  //         typeName: values.typeName,
+  //         testResult: values.testResult || "",
+  //         note: values.note || "",
+  //         status: values.status,
+  //         stage: relatedRoadmap?.stage || "", // Cung cấp giá trị mặc định nếu cần
+  //         stepNumber: relatedRoadmap?.stepNumber || 0, // Cung cấp giá trị mặc định nếu cần
+  //       },
+  //     ]);
 
-      message.success("Tạo mới hồ sơ bệnh án thành công");
-      setIsCreateMedicalDetailModalVisible(false);
-      medicalDetailForm.resetFields();
-    } catch (error) {
-      console.error("Lỗi khi tạo mới hồ sơ bệnh án:", error);
-      message.error("Tạo mới thất bại");
-    }
-  };
+  //     message.success("Tạo mới hồ sơ bệnh án thành công");
+  //     setIsCreateMedicalDetailModalVisible(false);
+  //     medicalDetailForm.resetFields();
+  //   } catch (error) {
+  //     console.error("Lỗi khi tạo mới hồ sơ bệnh án:", error);
+  //     message.error("Tạo mới thất bại");
+  //   }
+  // };
 
   //Consulation result
-  const handleConSulationR_typeTSubmit = (values: TestResultFormValues) => {
-    const newTrestmentResult: TypeTest = {
-      typeTestId: Date.now(),
-      name: values.name,
-      description: values.description,
-    };
-    setTestResults([...testResults, newTrestmentResult]);
-    setIsTestModalVisible(false);
-    treatmentResultForm.resetFields();
-  };
 
-  const mappedTreatmentResults = treatmentResult_typeTest.map((item) => ({
-    Treatment_result_ID: String(item.treatmentResultId),
-    Road_ID: String(item.treatmentRoadmapId),
-    Date: item.dateTreatmentResult,
-    Description: item.description,
-    Result: item.result,
-  }));
+  // const handleConSulationR_typeTSubmit = (values: TestResultFormValues) => {
+  //   const newTrestmentResult: TypeTest = {
+  //     typeTestId: Date.now(),
+  //     name: values.name,
+  //     description: values.description,
+  //   };
+  //   setTestResults([...testResults, newTrestmentResult]);
+  //   setIsTestModalVisible(false);
+  //   treatmentResultForm.resetFields();
+  // };
+
+  // const mappedTreatmentResults = treatmentResult_typeTest.map((item) => ({
+  //   Treatment_result_ID: String(item.treatmentResultId),
+  //   Road_ID: String(item.treatmentRoadmapId),
+  //   Date: item.dateTreatmentResult,
+  //   Description: item.description,
+  //   Result: item.result,
+  // }));
 
   const PatientDetailContent = () => (
     <div>
@@ -439,12 +439,12 @@ export default function PatientDetailPage() {
                   medicalRecordId={selectedMedicalRecordId}
                   // treatmentRoadmap={treatmentRoadmap}
                   // treatmentResults={treatmentResult_typeTest}
-                  medicalRecordDetails={medicalRecordDetails}
-                  consulationResults={consulationResult_typeTest}
-                  onAddTest={() => setIsTestModalVisible(true)}
+                  // medicalRecordDetails={medicalRecordDetails}
+                  // consulationResults={consulationResult_typeTest}
+                  // onAddTest={() => setIsTestModalVisible(true)}
                   // onUpdateTreatmentResult={showUpdateTreatmentResultModal}
-                  onUpdateDetail={showUpdateMedicalDetailModal}
-                  onAddDetail={showCreateMedicalDetailModal}
+                  // onUpdateDetail={showUpdateMedicalDetailModal}
+                  // onAddDetail={showCreateMedicalDetailModal}
                   // onAddTreatmentResult={() => {
                   //   setIsUpdateTreatmentResultModalVisible(true);
                   //   setEditingTreatmentResult(null);
@@ -493,7 +493,7 @@ export default function PatientDetailPage() {
                 )} */}
 
                 {/* Medical record detail */}
-                <MedicalDetailModal
+                {/* <MedicalDetailModal
                   open={isUpdateMedicalDetailModalVisible} // Sử dụng state đúng
                   onCancel={() => setIsUpdateMedicalDetailModalVisible(false)}
                   onSubmit={(values) => {
@@ -520,17 +520,17 @@ export default function PatientDetailPage() {
                   treatmentResults={treatmentResult_typeTest}
                   form={medicalDetailForm}
                   isEditing={false}
-                />
+                /> */}
 
                 {/* Consulation result - type test */}
-                <TestResultModal
+                {/* <TestResultModal
                   visible={isTestModalVisible}
                   onCancel={() => setIsTestModalVisible(false)}
                   onSubmit={handleConSulationR_typeTSubmit}
                   treatmentResults={mappedTreatmentResults}
                   medicalRecordDetails={medicalRecordDetails}
                   form={treatmentResultForm}
-                />
+                /> */}
               </>
             ),
           },
